Trait module_support::homa::HomaSubAccountXcm

pub trait HomaSubAccountXcm<AccountId, Balance> {
    fn transfer_staking_to_sub_account(
        sender: &AccountId,
        sub_account_index: u16,
        amount: Balance
    ) -> DispatchResult;
    fn withdraw_unbonded_from_sub_account(
        sub_account_index: u16,
        amount: Balance
    ) -> DispatchResult;
    fn bond_extra_on_sub_account(
        sub_account_index: u16,
        amount: Balance
    ) -> DispatchResult;
    fn unbond_on_sub_account(
        sub_account_index: u16,
        amount: Balance
    ) -> DispatchResult;
    fn get_xcm_transfer_fee() -> Balance;
    fn get_parachain_fee(location: MultiLocation) -> Balance;
}

Required Methods

fn transfer_staking_to_sub_account(
sender: &AccountId,
sub_account_index: u16,
amount: Balance
) -> DispatchResult

Cross-chain transfer staking currency to sub account on relaychain.

fn withdraw_unbonded_from_sub_account(
sub_account_index: u16,
amount: Balance
) -> DispatchResult

Send XCM message to the relaychain for sub account to withdraw_unbonded staking currency and send it back.

fn bond_extra_on_sub_account(
sub_account_index: u16,
amount: Balance
) -> DispatchResult

Send XCM message to the relaychain for sub account to bond extra.

fn unbond_on_sub_account(
sub_account_index: u16,
amount: Balance
) -> DispatchResult

Send XCM message to the relaychain for sub account to unbond.

fn get_xcm_transfer_fee() -> Balance

The fee of cross-chain transfer is deducted from the recipient.

fn get_parachain_fee(location: MultiLocation) -> Balance

The fee of parachain