Trait module_support::honzon::CDPTreasury

pub trait CDPTreasury<AccountId> {
    type Balance;
    type CurrencyId;

    fn get_surplus_pool() -> Self::Balance;
    fn get_debit_pool() -> Self::Balance;
    fn get_total_collaterals(id: Self::CurrencyId) -> Self::Balance;
    fn get_debit_proportion(amount: Self::Balance) -> Ratio;
    fn on_system_debit(amount: Self::Balance) -> DispatchResult;
    fn on_system_surplus(amount: Self::Balance) -> DispatchResult;
    fn issue_debit(
        who: &AccountId,
        debit: Self::Balance,
        backed: bool
    ) -> DispatchResult;
    fn burn_debit(who: &AccountId, debit: Self::Balance) -> DispatchResult;
    fn deposit_surplus(
        from: &AccountId,
        surplus: Self::Balance
    ) -> DispatchResult;
    fn withdraw_surplus(to: &AccountId, surplus: Self::Balance) -> DispatchResult;
    fn deposit_collateral(
        from: &AccountId,
        currency_id: Self::CurrencyId,
        amount: Self::Balance
    ) -> DispatchResult;
    fn withdraw_collateral(
        to: &AccountId,
        currency_id: Self::CurrencyId,
        amount: Self::Balance
    ) -> DispatchResult;
}

An abstraction of cdp treasury for Honzon Protocol.

Required Associated Types

type Balance

type CurrencyId

Required Methods

fn get_surplus_pool() -> Self::Balance

get surplus amount of cdp treasury

fn get_debit_pool() -> Self::Balance

get debit amount of cdp treasury

fn get_total_collaterals(id: Self::CurrencyId) -> Self::Balance

get collateral assets amount of cdp treasury

fn get_debit_proportion(amount: Self::Balance) -> Ratio

calculate the proportion of specific debit amount for the whole system

fn on_system_debit(amount: Self::Balance) -> DispatchResult

issue debit for cdp treasury

fn on_system_surplus(amount: Self::Balance) -> DispatchResult

issue surplus(stable currency) for cdp treasury

fn issue_debit(
who: &AccountId,
debit: Self::Balance,
backed: bool
) -> DispatchResult

issue debit to who if backed flag is true, means the debit to issue is backed on some assets, otherwise will increase same amount of debit to system debit.

fn burn_debit(who: &AccountId, debit: Self::Balance) -> DispatchResult

burn debit(stable currency) of who

fn deposit_surplus(from: &AccountId, surplus: Self::Balance) -> DispatchResult

deposit surplus(stable currency) to cdp treasury by from

fn withdraw_surplus(to: &AccountId, surplus: Self::Balance) -> DispatchResult

withdraw surplus(stable currency) from cdp treasury to to

fn deposit_collateral(
from: &AccountId,
currency_id: Self::CurrencyId,
amount: Self::Balance
) -> DispatchResult

deposit collateral assets to cdp treasury by who

fn withdraw_collateral(
to: &AccountId,
currency_id: Self::CurrencyId,
amount: Self::Balance
) -> DispatchResult

withdraw collateral assets of cdp treasury to who